+++ plume/src/main/kotlin/io/github/plume/oss/util/ExtractorConst.kt
@@ -2,7 +2,7 @@ package io.github.plume.oss.util
 
 object ExtractorConst {
     const val LANGUAGE_FRONTEND = "Plume"
-    const val LANGUAGE_FRONTEND_VERSION = "0.1"
+    const val LANGUAGE_FRONTEND_VERSION = "0.2.1"
     val PRIMITIVES = listOf("boolean", "byte", "char", "double", "float", "int", "long", "null", "short", "void")
     // block bodies
     const val ENTRYPOINT = "BODY"